🎨(style) 아래 로고 제거
