+++ Sedam/Sedam/Presentation/Post/PostMainView.swift
@@ -11,6 +11,8 @@ struct PostMainView: View {
     @EnvironmentObject private var router: Router
     @EnvironmentObject var viewModel: PostViewModel
     @EnvironmentObject var userViewModel: UserViewModel
+    @State private var date: Date = .now
+    @State private var isCalendarPresented = false
     
     private var indexedPosts: [(offset: Int, element: PostDTO)] {
         Array(viewModel.postList.enumerated())
@@ -26,7 +28,29 @@ struct PostMainView: View {
                 // 2) 화면 전체 크기를 차지시키되, 내용은 위쪽 정렬
                 VStack(spacing: 16) {
                     // 3) safe area 높이만큼 패딩
-                    ThemeView()
+                    ThemeView(
+                        date: $date,
+                        words: $viewModel.todayWords,
+                        wordsFont: .danjoBold18
+                    )
+                    
+                    HStack {
+                        Button {
+                            isCalendarPresented.toggle()
+                        } label: {
+                            Image(systemName: "calendar")
+                        }
+                        
+                        Button {
+                            viewModel.togglePostListType()
+                        } label : {
+                            if viewModel.postListType.sort == .likes {
+                                Image(systemName: "list.number")
+                            } else {
+                                Image(systemName: "clock")
+                            }
+                        }
+                    }
                     
                     // 4) 남은 공간은 스크롤뷰로 채움
                     ScrollView {
@@ -48,7 +72,26 @@ struct PostMainView: View {
                     alignment: .top     // VStack 전체를 위쪽 정렬
                 )
                 .padding(.top, proxy.safeAreaInsets.top + 48)
+                .onChange(of: date) { _, newDate in
+                    date = newDate
+                    viewModel.getTodayWords(by: date)
+                    viewModel.fetchPostList(sortBy: .likes, order: .desc, date: date)
+                }
+                .sheet(isPresented: $isCalendarPresented) {
+                    VStack {
+                        DatePicker ( "", selection: $date, in: ...Date(), displayedComponents: .date)
+                            .datePickerStyle(.graphical)
+                            .labelsHidden()
+                        
+                        Button("선택 완료") {
+                            isCalendarPresented = false
+                        }
+                        .padding(.top, 16)
+                        .font(.pretendardSemiBold14)
+                    }
+                }
                 .task {
+                    self.date = viewModel.postListType.startDate
                     viewModel.fetchPostList(sortBy: viewModel.postListType.sort, order: viewModel.postListType.order, date: viewModel.postListType.startDate)
                     userViewModel.fetchUserNickname()
                 }
