+++ Sedam/Sedam/Presentation/Post/PostUpdateView.swift
@@ -36,7 +36,6 @@ struct PostUpdateView: View {
                     .font(.maruburiRegular24)
                     .multilineTextAlignment(.center)
                     .padding(.vertical, 8)
-                LogoView(size: 10)
                 VStack {
                     ZStack(alignment: .topLeading) {
                         TextEditor(text: $content)
